refactor(view): put components and containers files into specific folder
create a container and a component folder and put into the corresponding files
